+++ Sources/AppIconKit/AlternateAppIcon.swift
@@ -11,6 +11,10 @@ import SwiftUI
 /// This model represents an alternate app icon, wich can be
 /// set to both list and set alternate app icons.
 ///
+/// The `icon` is the icon asset to display, while `iconName`
+/// is the name of the icon to set when the icon is selected.
+/// You can make set `iconName` to nil to reset the app icon.
+///
 /// The type is also a namespace of related icon types, like
 /// ``AlternateAppIcon/Item`` & ``AlternateAppIcon/ItemStyle``.
 public struct AlternateAppIcon {
@@ -19,10 +23,10 @@ public struct AlternateAppIcon {
     ///
     /// - Parameters:
     ///   - icon: The icon asset to display.
-    ///   - iconName: The name of the alternate icon.
+    ///   - iconName: The name of the alternate icon, if any.
     public init(
         icon: Image,
-        iconName: String
+        iconName: String?
     ) {
         self.icon = icon
         self.iconName = iconName
@@ -31,8 +35,8 @@ public struct AlternateAppIcon {
     /// The icon asset to display.
     public let icon: Image
 
-    /// The name of the alternate icon.
-    public let iconName: String
+    /// The name of the alternate icon, if any.
+    public let iconName: String?
 }
 
 public extension AlternateAppIcon {